pyproject.toml
[tool.poetry]
name = "chatplatform"
version = "0.1.0"
description = ""
authors = ["Ilja Grigorjev <extern.grigorjev_ilja@allianz-warranty.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.2.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
dataclasses-json = "0.6.4"
deprecated = "1.2.14"
dirtyjson = "1.0.8"
distro = "1.9.0"
exceptiongroup = "1.2.0"
fastapi = "0.109.2"
filelock = "3.13.1"
frozenlist = "1.4.1"
fsspec = "2024.2.0"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.3"
httpx = "0.26.0"
huggingface-hub = "0.20.3"
idna = "3.6"
jinja2 = "3.1.3"
joblib = "1.3.2"
llama-index-core = "^0.10.6.post1"
llama-index-embeddings-huggingface = "^0.1.1"
llama-index-llms-openai = "^0.1.2"
llama-index-llms-replicate = "^0.1.1"
markupsafe = "2.1.5"
marshmallow = "3.20.2"
mpmath = "1.3.0"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
networkx = "3.2.1"
nltk = "3.8.1"
numpy = "1.26.4"
openai = "1.12.0"
packaging = "23.2"
pandas = "2.2.0"
pillow = "10.2.0"
pydantic = "2.6.1"
pydantic-core = "2.16.2"
python-dateutil = "2.8.2"
pytz = "2024.1"
pyyaml = "6.0.1"
regex = "2023.12.25"
requests = "2.31.0"
safetensors = "0.4.2"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.27"
starlette = "0.36.3"
sympy = "1.12"
tenacity = "8.2.3"
tiktoken = "0.6.0"
tokenizers = "0.15.2"
torch = "2.2.0"
tqdm = "4.66.2"
transformers = "4.37.2"
typing-inspect = "0.9.0"
typing-extensions = "4.9.0"
tzdata = "2024.1"
urllib3 = "2.2.0"
uvicorn = "0.27.1"
wrapt = "1.16.0"
yarl = "1.9.4"
passlib = "^1.7.4"
python-dotenv = "1.0.1"
alembic = "1.13.1"
psycopg2-binary = "2.9.9"
email-validator = "2.1.0"
python-jose = "3.3.0"
pycryptodome = "3.20.0"
aio_pika = "9.4.0"
chardet = "5.2.0"
fitz = "0.0.1.dev2"
docx = "^0.2.4"
pymupdf =  "1.23.22"
python-multipart = "0.0.9"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
